added line to readme
I may not be able to get back to this. However you are free to use as long as UTEP IEEE allows it. (let me know if you need to be added as a contributor in order to allow changes to libraries as I wont be able to manage any pull requests.